+++ resources/views/admin/users/home.blade.php
@@ -20,11 +20,22 @@
 	<table class="table">
 		<thead class="bg-light">
 			<tr class="text-center">
+				{{-- <th scope="col" class="border-0" width="5%">
+				</th> --}}
 				<th scope="col" class="border-0" width="10%">
 					<span>ID</span> 
 				</th>
-				<th scope="col" class="border-0" width="30%">
+				<th scope="col" class="border-0" width="40%">
 					<span>Username</span>
+				</th>
+				<th scope="col" class="border-0" width="5%">
+					<span>Status<br/>Count</span>
+				</th>
+				<th scope="col" class="border-0" width="5%">
+					<span>Followers<br/>Count</span>
+				</th>
+				<th scope="col" class="border-0" width="5%">
+					<span>Following<br/>Count</span>
 				</th>
 				<th scope="col" class="border-0" width="30%">
 					<span>Actions</span>
@@ -32,17 +43,28 @@
 			</tr>
 		</thead>
 		<tbody>
-			@foreach($users as $user)
+			@foreach($users as $key => $user)
 			@if($user->status == 'deleted')
 			<tr class="font-weight-bold text-center user-row">
+				{{-- <th scope="row">
+					<div class="custom-control custom-checkbox account-select-check">
+						<input type="checkbox" class="custom-control-input" disabled>
+						<label class="custom-control-label"></label>
+					</div>
+				</th> --}}
 				<th scope="row">
 					<span class="text-danger" class="text-monospace">{{$user->id}}</span>
 				</th>
 				<td class="text-left">
+					<img src="/storage/avatars/default.jpg" width="20" height="20" class="rounded-circle mr-1" />
+
 					<span title="{{$user->username}}" data-toggle="tooltip" data-placement="bottom">
 						<span class="text-danger">{{$user->username}}</span>
 					</span>
 				</td>
+				<td>0</td>
+				<td>0</td>
+				<td>0</td>
 				<td>
 					<span class="font-weight-bold small">
 						<span class="text-danger">Account Deleted</span>
@@ -51,16 +73,46 @@
 			</tr>
 			@else 
 			<tr class="font-weight-bold text-center user-row">
+				{{-- <th scope="row">
+					<div class="custom-control custom-checkbox account-select-check">
+						<input type="checkbox" id="{{$key}}" class="custom-control-input">
+						<label class="custom-control-label" for={{$key}}></label>
+					</div>
+				</th> --}}
 				<th scope="row">
 					<span class="text-monospace">{{$user->id}}</span>
 				</th>
-				<td class="text-left">
+				<td class="text-left d-flex align-items-center">
+					@if($user->account)
+					<img src="{{$user->account['avatar']}}" width="20" height="20" class="rounded-circle mr-2" />
+					@endif
 					<span title="{{$user->username}}" data-toggle="tooltip" data-placement="bottom">
 						<span>{{$user->username}}</span>
 						@if($user->is_admin)
 						<i class="text-danger fas fa-certificate" title="Admin"></i>
 						@endif
 					</span>
+				</td>
+				<td>
+					@if($user->account)
+					 {{$user->account['statuses_count']}}
+					@else
+					0
+					@endif
+				</td>
+				<td>
+					@if($user->account)
+					 {{$user->account['followers_count']}}
+					@else
+					0
+					@endif
+				</td>
+				<td>
+					@if($user->account)
+					 {{$user->account['following_count']}}
+					@else
+					0
+					@endif
 				</td>
 				<td>
 					<span class="action-row font-weight-lighter">
@@ -75,6 +127,11 @@
 						<a href="/i/admin/users/modtools/{{$user->id}}" class="pr-2 text-muted small font-weight-bold" title="Moderation Logs" data-toggle="tooltip" data-placement="bottom">
 							Mod Tools
 						</a>
+						@if($user->status !== 'deleted' && !$user->is_admin)
+						<a href="/i/admin/users/delete/{{$user->id}}" class="pr-2 text-muted small font-weight-bold" title="Delete account" data-toggle="tooltip" data-placement="bottom" onclick="deleteAccount({{$user->id}})">
+							Delete
+						</a>
+						@endif
 					</span>
 				</td>
 			</tr>
@@ -117,5 +174,20 @@
 			el.text(filesize(size, {round: 0}));
 		});
 	});
+
+	function deleteAccount(id) {
+		event.preventDefault();
+		if(!window.confirm('Are you sure you want to delete this account?')) {
+			return;
+		}
+
+		axios.post('/i/admin/users/delete/' + id)
+		.then(res => {
+			swal('Account Deleted', 'Successfully deleted this account! This page will refresh once you press OK.', 'success')
+			.then(res => {
+				window.location.reload();
+			});
+		})
+	}
 </script>
 @endpush